Rework geturls, worker, and clone_repo
It seems better to pass the config object itself rather than the
list of URLs. This way, the section name can be used as a directory
name and is not retrieved by its URL.

Futhermore, if I decide to add more options in the config file, it
can easily retrieved.